+++ api/Cargo.toml
@@ -24,3 +24,4 @@ openidconnect = "4.0.0-rc.1"
 url = "2.5.4"
 env_logger = "0.11.6"
 reqwest = "0.12.12"
+anyhow = "1.0.95"
