+++ api/src/handlers/open_id_connect/mod.rs
@@ -14,134 +14,12 @@
 //! ...and follow the instructions.
 //!
 
-use openidconnect::core::{
-    CoreAuthDisplay, CoreClaimName, CoreClaimType, CoreClient, CoreClientAuthMethod, CoreGrantType,
-    CoreIdTokenClaims, CoreIdTokenVerifier, CoreJsonWebKey, CoreJweContentEncryptionAlgorithm,
-    CoreJweKeyManagementAlgorithm, CoreResponseMode, CoreResponseType, CoreRevocableToken,
-    CoreSubjectIdentifierType,
-};
-use openidconnect::reqwest;
-use openidconnect::{
-    AdditionalProviderMetadata, AuthenticationFlow, AuthorizationCode, ClientId, ClientSecret,
-    CsrfToken, IssuerUrl, Nonce, OAuth2TokenResponse, ProviderMetadata, RedirectUrl, RevocationUrl,
-    Scope,
-};
-use serde::{Deserialize, Serialize};
-use url::Url;
-
-use std::env;
-use std::io::{BufRead, BufReader, Write};
-use std::net::TcpListener;
-use std::process::exit;
-use reqwest::ClientBuilder;
-
-fn handle_error<T: std::error::Error>(fail: &T, msg: &'static str) {
-    let mut err_msg = format!("ERROR: {}", msg);
-    let mut cur_fail: Option<&dyn std::error::Error> = Some(fail);
-    while let Some(cause) = cur_fail {
-        err_msg += &format!("\n    caused by: {}", cause);
-        cur_fail = cause.source();
-    }
-    println!("{}", err_msg);
-    exit(1);
-}
-
-// Teach openidconnect-rs about a Google custom extension to the OpenID Discovery response that we can use as the RFC
-// 7009 OAuth 2.0 Token Revocation endpoint. For more information about the Google specific Discovery response see the
-// Google OpenID Connect service documentation at: https://developers.google.com/identity/protocols/oauth2/openid-connect#discovery
-#[derive(Clone, Debug, Deserialize, Serialize)]
-struct RevocationEndpointProviderMetadata {
-    revocation_endpoint: String,
-}
-impl AdditionalProviderMetadata for RevocationEndpointProviderMetadata {}
-type GoogleProviderMetadata = ProviderMetadata<
-    RevocationEndpointProviderMetadata,
-    CoreAuthDisplay,
-    CoreClientAuthMethod,
-    CoreClaimName,
-    CoreClaimType,
-    CoreGrantType,
-    CoreJweContentEncryptionAlgorithm,
-    CoreJweKeyManagementAlgorithm,
-    CoreJsonWebKey,
-    CoreResponseMode,
-    CoreResponseType,
-    CoreSubjectIdentifierType,
->;
 
 pub async fn fjonk_e_bonke() {
     // get env vars
     env_logger::init();
 
-    let google_client_id = ClientId::new(
-        env::var("GOOGLE_CLIENT_ID").expect("Missing the GOOGLE_CLIENT_ID environment variable."),
-    );
-    let google_client_secret = ClientSecret::new(
-        env::var("GOOGLE_CLIENT_SECRET")
-            .expect("Missing the GOOGLE_CLIENT_SECRET environment variable."),
-    );
-    let issuer_url =
-        IssuerUrl::new("https://accounts.google.com".to_string()).unwrap_or_else(|err| {
-            handle_error(&err, "Invalid issuer URL");
-            unreachable!();
-        });
-
-    let http_client = ClientBuilder::new()
-        // Following redirects opens the client up to SSRF vulnerabilities.
-        .redirect(reqwest::redirect::Policy::none())
-        .build()
-        .unwrap_or_else(|err| {
-            handle_error(&err, "Failed to build HTTP client");
-            unreachable!();
-        });
-
-    // Fetch Google's OpenID Connect discovery document.
-    //
-    // Note: If we don't care about token revocation we can simply use CoreProviderMetadata here
-    // instead of GoogleProviderMetadata. If instead we wanted to optionally use the token
-    // revocation endpoint if it seems to be supported we could do something like this:
-    //   #[derive(Clone, Debug, Deserialize, Serialize)]
-    //   struct AllOtherProviderMetadata(HashMap<String, serde_json::Value>);
-    //   impl AdditionalClaims for AllOtherProviderMetadata {}
-    // And then test for the presence of "revocation_endpoint" in the map returned by a call to
-    // .additional_metadata().
-
-    let provider_metadata = GoogleProviderMetadata::discover_async(issuer_url, &http_client).await
-        .unwrap_or_else(|err| {
-            handle_error(&err, "Failed to discover OpenID Provider");
-            unreachable!();
-        });
-
-    let revocation_endpoint = provider_metadata
-        .additional_metadata()
-        .revocation_endpoint
-        .clone();
-    println!(
-        "Discovered Google revocation endpoint: {}",
-        revocation_endpoint
-    );
-
-    // Set up the config for the Google OAuth2 process.
-    let client = CoreClient::from_provider_metadata(
-        provider_metadata,
-        google_client_id,
-        Some(google_client_secret),
-    )
-        // This example will be running its own server at localhost:8080.
-        // See below for the server implementation.
-        .set_redirect_uri(
-            RedirectUrl::new("http://localhost:8080".to_string()).unwrap_or_else(|err| {
-                handle_error(&err, "Invalid redirect URL");
-                unreachable!();
-            }),
-        )
-        // Google supports OAuth 2.0 Token Revocation (RFC-7009)
-        .set_revocation_url(
-            RevocationUrl::new(revocation_endpoint).unwrap_or_else(|err| {
-                handle_error(&err, "Invalid revocation endpoint URL");
-                unreachable!();
-            }),
-        );
+    
 
     // Generate the authorization URL to which we'll redirect the user.
     let (authorize_url, csrf_state, nonce) = client
